// src/services/-admin-categories-post.schemas.ts
export interface I18nLocalizedMap {[key: string]: string}

export interface DtoCreateCategoryRequest {
  description?: string;
  descriptionI18n?: I18nLocalizedMap;
  is_active?: boolean;
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  nameI18n?: I18nLocalizedMap;
  parent_id?: number;
  slug?: string;
}